revert useless change and add comments
